requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
requests~=2.31.0
convert-numbers~=0.4
pillow~=10.1.0